+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.5.2
 
   - Added test for Fortigate FortiOS 5.2 (Netflow v9)
+  - Added permission check to templates cache (Issue #80)
   - Clarified confusing warning about missing templates
 
 ## 3.5.1